# demo_w7/flask-server/openapi_server/models/category.py
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model import Model
from openapi_server import util


class Category(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, id=None, name=None, slug=None, parent_id=None):  # noqa: E501
        """Category - a model defined in OpenAPI

        :param id: The id of this Category.  # noqa: E501
        :type id: int
        :param name: The name of this Category.  # noqa: E501
        :type name: str
        :param slug: The slug of this Category.  # noqa: E501
        :type slug: str
        :param parent_id: The parent_id of this Category.  # noqa: E501
        :type parent_id: int
        """
        self.openapi_types = {
            'id': int,
            'name': str,
            'slug': str,
            'parent_id': int
        }

        self.attribute_map = {
            'id': 'id',
            'name': 'name',
            'slug': 'slug',
            'parent_id': 'parent_id'
        }

        self._id = id
        self._name = name
        self._slug = slug
        self._parent_id = parent_id

    @classmethod
    def from_dict(cls, dikt) -> 'Category':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Category of this Category.  # noqa: E501
        :rtype: Category
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self) -> int:
        """Gets the id of this Category.


        :return: The id of this Category.
        :rtype: int
        """
        return self._id

    @id.setter
    def id(self, id: int):
        """Sets the id of this Category.


        :param id: The id of this Category.
        :type id: int
        """

        self._id = id

    @property
    def name(self) -> str:
        """Gets the name of this Category.


        :return: The name of this Category.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name: str):
        """Sets the name of this Category.


        :param name: The name of this Category.
        :type name: str
        """

        self._name = name

    @property
    def slug(self) -> str:
        """Gets the slug of this Category.


        :return: The slug of this Category.
        :rtype: str
        """
        return self._slug

    @slug.setter
    def slug(self, slug: str):
        """Sets the slug of this Category.


        :param slug: The slug of this Category.
        :type slug: str
        """

        self._slug = slug

    @property
    def parent_id(self) -> int:
        """Gets the parent_id of this Category.


        :return: The parent_id of this Category.
        :rtype: int
        """
        return self._parent_id

    @parent_id.setter
    def parent_id(self, parent_id: int):
        """Sets the parent_id of this Category.


        :param parent_id: The parent_id of this Category.
        :type parent_id: int
        """

        self._parent_id = parent_id