# demo_w7/flask-server/openapi_server/models/category_create.py
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model import Model
from openapi_server import util


class CategoryCreate(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, name=None, parent_id=None):  # noqa: E501
        """CategoryCreate - a model defined in OpenAPI

        :param name: The name of this CategoryCreate.  # noqa: E501
        :type name: str
        :param parent_id: The parent_id of this CategoryCreate.  # noqa: E501
        :type parent_id: int
        """
        self.openapi_types = {
            'name': str,
            'parent_id': int
        }

        self.attribute_map = {
            'name': 'name',
            'parent_id': 'parent_id'
        }

        self._name = name
        self._parent_id = parent_id

    @classmethod
    def from_dict(cls, dikt) -> 'CategoryCreate':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The CategoryCreate of this CategoryCreate.  # noqa: E501
        :rtype: CategoryCreate
        """
        return util.deserialize_model(dikt, cls)

    @property
    def name(self) -> str:
        """Gets the name of this CategoryCreate.


        :return: The name of this CategoryCreate.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name: str):
        """Sets the name of this CategoryCreate.


        :param name: The name of this CategoryCreate.
        :type name: str
        """
        if name is None:
            raise ValueError("Invalid value for `name`, must not be `None`")  # noqa: E501

        self._name = name

    @property
    def parent_id(self) -> int:
        """Gets the parent_id of this CategoryCreate.


        :return: The parent_id of this CategoryCreate.
        :rtype: int
        """
        return self._parent_id

    @parent_id.setter
    def parent_id(self, parent_id: int):
        """Sets the parent_id of this CategoryCreate.


        :param parent_id: The parent_id of this CategoryCreate.
        :type parent_id: int
        """

        self._parent_id = parent_id