# demo_w7/flask-server/openapi_server/models/product.py
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model import Model
from openapi_server import util


class Product(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, id=None, name=None, description=None, price=None, sku=None, category_id=None, stock_quantity=None, status=None, created_at=None):  # noqa: E501
        """Product - a model defined in OpenAPI

        :param id: The id of this Product.  # noqa: E501
        :type id: int
        :param name: The name of this Product.  # noqa: E501
        :type name: str
        :param description: The description of this Product.  # noqa: E501
        :type description: str
        :param price: The price of this Product.  # noqa: E501
        :type price: float
        :param sku: The sku of this Product.  # noqa: E501
        :type sku: str
        :param category_id: The category_id of this Product.  # noqa: E501
        :type category_id: int
        :param stock_quantity: The stock_quantity of this Product.  # noqa: E501
        :type stock_quantity: int
        :param status: The status of this Product.  # noqa: E501
        :type status: str
        :param created_at: The created_at of this Product.  # noqa: E501
        :type created_at: datetime
        """
        self.openapi_types = {
            'id': int,
            'name': str,
            'description': str,
            'price': float,
            'sku': str,
            'category_id': int,
            'stock_quantity': int,
            'status': str,
            'created_at': datetime
        }

        self.attribute_map = {
            'id': 'id',
            'name': 'name',
            'description': 'description',
            'price': 'price',
            'sku': 'sku',
            'category_id': 'category_id',
            'stock_quantity': 'stock_quantity',
            'status': 'status',
            'created_at': 'created_at'
        }

        self._id = id
        self._name = name
        self._description = description
        self._price = price
        self._sku = sku
        self._category_id = category_id
        self._stock_quantity = stock_quantity
        self._status = status
        self._created_at = created_at

    @classmethod
    def from_dict(cls, dikt) -> 'Product':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Product of this Product.  # noqa: E501
        :rtype: Product
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self) -> int:
        """Gets the id of this Product.


        :return: The id of this Product.
        :rtype: int
        """
        return self._id

    @id.setter
    def id(self, id: int):
        """Sets the id of this Product.


        :param id: The id of this Product.
        :type id: int
        """

        self._id = id

    @property
    def name(self) -> str:
        """Gets the name of this Product.


        :return: The name of this Product.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name: str):
        """Sets the name of this Product.


        :param name: The name of this Product.
        :type name: str
        """

        self._name = name

    @property
    def description(self) -> str:
        """Gets the description of this Product.


        :return: The description of this Product.
        :rtype: str
        """
        return self._description

    @description.setter
    def description(self, description: str):
        """Sets the description of this Product.


        :param description: The description of this Product.
        :type description: str
        """

        self._description = description

    @property
    def price(self) -> float:
        """Gets the price of this Product.


        :return: The price of this Product.
        :rtype: float
        """
        return self._price

    @price.setter
    def price(self, price: float):
        """Sets the price of this Product.


        :param price: The price of this Product.
        :type price: float
        """

        self._price = price

    @property
    def sku(self) -> str:
        """Gets the sku of this Product.


        :return: The sku of this Product.
        :rtype: str
        """
        return self._sku

    @sku.setter
    def sku(self, sku: str):
        """Sets the sku of this Product.


        :param sku: The sku of this Product.
        :type sku: str
        """

        self._sku = sku

    @property
    def category_id(self) -> int:
        """Gets the category_id of this Product.


        :return: The category_id of this Product.
        :rtype: int
        """
        return self._category_id

    @category_id.setter
    def category_id(self, category_id: int):
        """Sets the category_id of this Product.


        :param category_id: The category_id of this Product.
        :type category_id: int
        """

        self._category_id = category_id

    @property
    def stock_quantity(self) -> int:
        """Gets the stock_quantity of this Product.


        :return: The stock_quantity of this Product.
        :rtype: int
        """
        return self._stock_quantity

    @stock_quantity.setter
    def stock_quantity(self, stock_quantity: int):
        """Sets the stock_quantity of this Product.


        :param stock_quantity: The stock_quantity of this Product.
        :type stock_quantity: int
        """

        self._stock_quantity = stock_quantity

    @property
    def status(self) -> str:
        """Gets the status of this Product.


        :return: The status of this Product.
        :rtype: str
        """
        return self._status

    @status.setter
    def status(self, status: str):
        """Sets the status of this Product.


        :param status: The status of this Product.
        :type status: str
        """

        self._status = status

    @property
    def created_at(self) -> datetime:
        """Gets the created_at of this Product.


        :return: The created_at of this Product.
        :rtype: datetime
        """
        return self._created_at

    @created_at.setter
    def created_at(self, created_at: datetime):
        """Sets the created_at of this Product.


        :param created_at: The created_at of this Product.
        :type created_at: datetime
        """

        self._created_at = created_at